# services/core-go/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.CashflowEntry:
    properties:
      act_explaination:
        type: string
      act_information:
        type: string
      created_at:
        type: string
      created_by:
        type: string
      date_of_entry:
        type: string
      debit:
        type: number
      due_date:
        type: string
      entry_type:
        $ref: '#/definitions/domain.EntryType'
      grand_cost:
        type: number
      grand_selling:
        type: number
      id:
        type: integer
      kredit:
        type: number
      margin_pct:
        type: number
      profit:
        type: number
      remarks:
        $ref: '#/definitions/domain.PaymentStatus'
      saldo:
        type: number
      sequence_no:
        type: integer
      top_days:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
      vendor_id:
        type: integer
      vendor_name_raw:
        type: string
    type: object
...
  domain.EntryType:
    enum:
    - SHIPMENT
    - TOP_UP
    type: string
    x-enum-varnames:
    - EntryShipment
    - EntryTopUp
  domain.PaymentStatus:
    enum:
    - PAID
    - UNPAID
    - PENDING
    type: string
    x-enum-varnames:
    - PaymentPaid
    - PaymentUnpaid
    - PaymentPending
...
  handler.UpdateStatusRequest:
    properties:
      remarks:
        $ref: '#/definitions/domain.PaymentStatus'
    type: object
  response.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: boolean
    type: object
  response.PaginationMeta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    High Performance Backend Service for PT. Adijayantara Logistics Indonesia.
    Architecture & Headers:
    - X-API-Version: Target API versioning (e.g. 'v1', 'v2')
    - X-Idempotency-Key: Unique UUID token for safe mutations (POST/PUT/PATCH/DELETE) to prevent duplicates
    - X-Request-ID: Distributed tracing correlation ID
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Cashflow & Shipment Management API
  version: "1.0"
paths:
...
  /cashflow:
    get:
      parameters:
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: limit
        type: integer
      - default: ASC
        description: Sort direction (ASC/DESC)
        in: query
        name: sort
        type: string
      - description: Start date (YYYY-MM-DD)
        in: query
        name: date_from
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: date_to
        type: string
      - description: Entry type (SHIPMENT/TOP_UP)
        in: query
        name: entry_type
        type: string
      - description: Payment status (PAID/UNPAID/PENDING)
        in: query
        name: remarks
        type: string
      - description: Vendor search keyword
        in: query
        name: vendor_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard cashflow entries
      tags:
      - cashflow
  /cashflow/{id}:
    delete:
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a cashflow entry
      tags:
      - cashflow
    put:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Entry Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CashflowEntry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update a cashflow entry
      tags:
      - cashflow
  /cashflow/{id}/status:
    patch:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Payment Status Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Update payment status (Remarks)
      tags:
      - cashflow
  /cashflow/export:
    get:
      produces:
      - application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: Export cashflow data to Excel
      tags:
      - cashflow
  /cashflow/import:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Excel File
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Import cashflow data from Excel
      tags:
      - cashflow
  /cashflow/shipment:
    post:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Shipment Entry Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CashflowEntry'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Record a Shipment entry
      tags:
      - cashflow
  /cashflow/summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Get dashboard cashflow summary
      tags:
      - cashflow
  /cashflow/topup:
    post:
      consumes:
      - application/json
      parameters:
      - default: a0eebc99-9c0b-4ef8-bb6d-6bb9bd380a11
        description: Idempotency Key (UUID unik pencegah duplikasi transaksi)
        in: header
        name: X-Idempotency-Key
        type: string
      - description: 'API Version (default: v1)'
        in: header
        name: X-API-Version
        type: string
      - description: Top-Up Entry Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.CashflowEntry'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.APIResponse'
      security:
      - BearerAuth: []
      summary: Record a Top-Up entry
      tags:
      - cashflow
...
securityDefinitions:
  ApiVersion:
    description: Target API Version Routing (e.g. v1, v2)
    in: header
    name: X-API-Version
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and the session token.
    in: header
    name: Authorization
    type: apiKey
  IdempotencyKey:
    description: Unique UUID idempotency key for mutation operations (POST/PUT/PATCH/DELETE)
    in: header
    name: X-Idempotency-Key
    type: apiKey
swagger: "2.0"